add stats to django admin, endpoint to delete today's data from db

+++ backend/scrapy/tychy/flat_schema.py
@@ -15,7 +15,7 @@ class FlatSchema(BaseModel):
     rent: Optional[int] = None
 
     rooms: Optional[int] = None
-    space: Optional[int] = None
+    space: Optional[float] = None
     status: Optional[str]
     # floor : str = None
     ownership: Optional[str] = None
